requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
boto3==1.34.162
python-dotenv==1.0.1
pyserial==3.5
pynmea2==1.19.0
rplidar-roboticia==0.9.5
smbus2==0.4.3
requests==2.31.0 